requirements.txt
groq>=0.4.0
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
pytest>=7.4.3
